test fix

+++ here-naksha-lib-core/src/test/java/com/here/naksha/lib/core/models/storage/XyzFeatureCodecTest.java
@@ -46,7 +46,7 @@ public class XyzFeatureCodecTest extends SessionTest {
         BigInt64(now.toEpochMilli()),
         BigInt64(now.toEpochMilli()),
         BigInt64(2),
-        0,
+        (short) 0,
         1,
         BigInt64(now.toEpochMilli()),
         BigInt64(3),
